legacy/requirements.txt
scrapling
python-dotenv
msgspec
curl_cffi
patchright
browserforge
schedule
